Create Vehicle.java
Vehicle class is our base class and In vehicle class I had used one parameterized constructor only for chasisNo which is required while when JVM had to check the similar chasisNo. And equals method from Object class is override as equals method in object class compares reference and we want comparision of content , so I overrided it and then downcasted to Compare the content in the Chasis No variable.