Create Tester.java
In this code few methods of arraylist is tested such as "add method" which is used for append and also for inserting Integer as per position, then "contains method" is used for searching element in array, "remove method " is used to delete the element in array, I have also shown that how to double the value greater than 50 in arraylist, as it is index based loop so there I used for loop rather than for each loop and then for replacing the number the "set method" is used and to get the element "get method" is used and then it is doubled. Next sorting of elements is done as arraylist is ordered list not a sorted list so by using collections class we did sorting of arraylist in ascending order.